fix: order of types in list

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -4,7 +4,7 @@
 import typing
 
 
-def sum_mixed_list(mxd_lst: typing.List[typing.Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: typing.List[typing.Union[int, float]]) -> float:
     """adds list of floats together function"""
     result: float = 0
